+++ doc/release/RELEASE-NOTES.md
@@ -13,9 +13,13 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 #### Migrations
 * 0179_introduce_enable_v1_claims_flow.py - This migration introduces a new `enable_v1_claims_flow` feature flag that allows switching to the v1 claims flow.
+* 0180_add_unit_to_partner_field.py - This migration added new field `unit` to `PartnerField` model.
 
 ### What's new
 * [OSDEV-2176](https://opensupplyhub.atlassian.net/browse/OSDEV-2176) - Added feature flag for v1 claims flow.
+* [OSDEV-2065](https://opensupplyhub.atlassian.net/browse/OSDEV-2065) - Updated v1 production locations `POST/PATCH` endpoints to include partner fields:
+    * Added `unit` field to `PartnerField` model
+    * Added type validation for submitted partner fields
 
 ### Release instructions
 * Ensure that the following commands are included in the `post_deployment` command: